bot, plugins, tools: remove sopel.tools.jobs
