+++ test/plugins/test_plugins_jobs.py
@@ -1,12 +1,19 @@
-"""Tests for Job Scheduler"""
+"""Tests for the ``sopel.plugins.jobs`` module."""
 from __future__ import annotations
 
 import time
+import typing
 
 import pytest
 
 from sopel import plugin
-from sopel.tools import jobs
+from sopel.plugins import jobs
+
+
+if typing.TYPE_CHECKING:
+    from sopel.bot import Sopel
+    from sopel.config import Config
+    from sopel.tests.factories import BotFactory, ConfigFactory
 
 
 TMP_CONFIG = """
@@ -22,11 +29,16 @@ class WithJobMockException(Exception):
 
 
 @pytest.fixture
-def mockconfig(configfactory):
+def mockconfig(configfactory: ConfigFactory) -> Config:
     return configfactory('config.cfg', TMP_CONFIG)
 
 
-def test_jobscheduler_stop(mockconfig, botfactory):
+@plugin.interval(5)
+def mock_job(bot: Sopel) -> None:
+    pass
+
+
+def test_jobscheduler_stop(mockconfig: Config, botfactory: BotFactory) -> None:
     mockbot = botfactory(mockconfig)
     scheduler = jobs.Scheduler(mockbot)
     assert not scheduler.stopping.is_set(), 'Stopping must not be set at init'
@@ -37,65 +49,59 @@ def test_jobscheduler_stop(mockconfig, botfactory):
 
 def test_job_is_ready_to_run():
     now = time.time()
-    job = jobs.Job([5])
+    job = jobs.Job(
+        [5],
+        plugin=mock_job.plugin_name,
+        label=mock_job.label,
+        handler=mock_job,
+    )
 
     assert job.is_ready_to_run(now + 20)
     assert not job.is_ready_to_run(now - 20)
 
 
 def test_job_str():
-    job = jobs.Job([5])
-    expected = '<Job (unknown) [5s]>'
+    job = jobs.Job(
+        [5],
+        plugin=None,
+        label=mock_job.label,
+        handler=mock_job,
+    )
+    expected = '<Job __anonymous_plugin__.mock_job [5s]>'
 
     assert str(job) == expected
 
 
 def test_job_str_intervals():
-    job = jobs.Job([5, 60, 30])
-    expected = '<Job (unknown) [5s, 30s, 60s]>'
+    job = jobs.Job(
+        [5, 60, 30],
+        plugin=None,
+        label=mock_job.label,
+        handler=mock_job,
+    )
+    expected = '<Job __anonymous_plugin__.mock_job [5s, 30s, 60s]>'
 
     assert str(job) == expected
 
 
-def test_job_str_handler():
-    def handler():
-        pass
-
-    job = jobs.Job([5], handler=handler)
-    expected = '<Job handler [5s]>'
-
-    assert str(job) == expected
-
-
-def test_job_str_handler_plugin():
-    def handler():
-        pass
-
-    job = jobs.Job([5], plugin='testplugin', handler=handler)
-    expected = '<Job testplugin.handler [5s]>'
-
-    assert str(job) == expected
-
-
-def test_job_str_label():
-    def handler():
-        pass
-
-    job = jobs.Job([5], label='testhandler', handler=handler)
-    expected = '<Job testhandler [5s]>'
+def test_job_str_plugin():
+    job = jobs.Job(
+        [5],
+        plugin='testplugin',
+        label=mock_job.label,
+        handler=mock_job,
+    )
+    expected = '<Job testplugin.mock_job [5s]>'
 
     assert str(job) == expected
 
 
 def test_job_str_label_plugin():
-    def handler():
-        pass
-
     job = jobs.Job(
         [5],
         label='testhandler',
         plugin='testplugin',
-        handler=handler,
+        handler=mock_job,
     )
     expected = '<Job testplugin.testhandler [5s]>'
 
@@ -103,14 +109,11 @@ def test_job_str_label_plugin():
 
 
 def test_job_str_label_plugin_intervals():
-    def handler():
-        pass
-
     job = jobs.Job(
         [5, 3600, 60],
         label='testhandler',
         plugin='testplugin',
-        handler=handler,
+        handler=mock_job,
     )
     expected = '<Job testplugin.testhandler [5s, 60s, 3600s]>'
 
@@ -119,7 +122,12 @@ def test_job_str_label_plugin_intervals():
 
 def test_job_next():
     timestamp = 523549800
-    job = jobs.Job([5])
+    job = jobs.Job(
+        [5],
+        label=mock_job.label,
+        plugin=mock_job.plugin_name,
+        handler=mock_job,
+    )
     job.next_times[5] = timestamp
 
     job.next(timestamp)
@@ -153,7 +161,12 @@ def test_job_next():
 
 def test_job_next_many():
     timestamp = 523549800
-    job = jobs.Job([5, 30])
+    job = jobs.Job(
+        [5, 30],
+        label=mock_job.label,
+        plugin=mock_job.plugin_name,
+        handler=mock_job,
+    )
     job.next_times[5] = timestamp + 5
     job.next_times[30] = timestamp + 30
 
@@ -201,7 +214,7 @@ def test_job_next_many():
 def test_job_from_callable(mockconfig):
     @plugin.interval(5)
     @plugin.label('testjob')
-    def handler(manager):
+    def handler(bot: Sopel):
         """The job's docstring."""
         return 'tested'
 
@@ -222,7 +235,12 @@ def test_job_from_callable(mockconfig):
 
 
 def test_job_with():
-    job = jobs.Job([5])
+    job = jobs.Job(
+        [5],
+        label=mock_job.label,
+        plugin=mock_job.plugin_name,
+        handler=mock_job,
+    )
     # play with time: move 1s back in the future
     last_time = job.next_times[5] = time.time() - 1
 
@@ -238,7 +256,12 @@ def test_job_with():
 
 
 def test_job_with_exception():
-    job = jobs.Job([5])
+    job = jobs.Job(
+        [5],
+        label=mock_job.label,
+        plugin=mock_job.plugin_name,
+        handler=mock_job,
+    )
     # play with time: move 1s back in the future
     last_time = job.next_times[5] = time.time() - 1
